Refactor to pkgs and move to godep.
