+++ Dockerfile
@@ -1,11 +1,16 @@
-FROM golang:1.11 as test
-WORKDIR /app
-COPY go.* /app/
-RUN go mod download
-COPY *.go /app/
-RUN CGO_ENABLED=0 go build -mod=readonly -o /shell-operator ./...
+FROM golang:1.11.0 as base
+RUN go get -u github.com/golang/dep/cmd/dep
+WORKDIR /go/src/github.com/MYOB-Technology/shell-operator
+
+FROM golang:1.11.0 as build
+WORKDIR /go/src/github.com/MYOB-Technology/shell-operator
+COPY --from=base /go/bin/dep /go/bin/dep
+COPY Gopkg.* /go/src/github.com/MYOB-Technology/shell-operator/
+RUN dep ensure -v -vendor-only
+COPY . /go/src/github.com/MYOB-Technology/shell-operator/
+RUN CGO_ENABLED=0 go build -o /shell-operator main.go
 
 FROM scratch
 COPY example/shell-conf.yaml /app/shell-config.yaml
-COPY --from=test /shell-operator /shell-operator
+COPY --from=build /shell-operator /shell-operator
 ENTRYPOINT ["/shell-operator"]
