+++ main.go
@@ -5,6 +5,8 @@ import (
 	"log"
 	"os"
 
+	"github.com/MYOB-Technology/shell-operator/pkg/watcher"
+
 	"sigs.k8s.io/controller-runtime/pkg/client/config"
 	"sigs.k8s.io/controller-runtime/pkg/manager"
 	"sigs.k8s.io/controller-runtime/pkg/runtime/signals"
@@ -24,7 +26,7 @@ func main() {
 		log.Fatal(err)
 	}
 
-	shellConfig, errs := ParseAndValidateConfig(confIn)
+	shellConfig, errs := watcher.ParseAndValidateConfig(confIn)
 
 	if len(errs) > 0 {
 		log.Fatal(errs)
@@ -42,7 +44,7 @@ func main() {
 		log.Fatal(err)
 	}
 
-	err = SetupWatches(mgr, shellConfig)
+	err = watcher.SetupWatches(mgr, shellConfig)
 
 	if err != nil {
 		log.Fatal(err)
