+++ pkg/watcher/config.go
@@ -1,4 +1,4 @@
-package main
+package watcher
 
 import (
 	"io"
