+++ pkg/watcher/controller.go
@@ -1,4 +1,4 @@
-package main
+package watcher
 
 import (
 	"strings"
