+++ pkg/watcher/controller_test.go
@@ -1,4 +1,4 @@
-package main
+package watcher
 
 import (
 	"testing"
