+++ pkg/watcher/dynamic.go
@@ -1,4 +1,4 @@
-package main
+package watcher
 
 import (
 	"fmt"
