+++ pkg/watcher/dynamic_test.go
@@ -1,4 +1,4 @@
-package main
+package watcher
 
 import (
 	"reflect"
